refactor: Answer를 Survey에서 분리(#102)
